requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
pytest==7.2.0
pytest-xdist==3.1.0
requests==2.28.1
datetime==4.8
jsonschema==4.17.3
